refactor player weapon system to have player and bullet weapons
