+++ Assets/Scripts/Weapons/BulletWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e51e37c15ed4bb64c97d107d450eb81f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
