+++ Assets/Scripts/Weapons/PlayerWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 418769f1a024dd94f88f5f905b0c9f61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
